+++ weblate/accounts/models.py
@@ -37,7 +37,7 @@ from south.signals import post_migrate
 from social.apps.django_app.default.models import UserSocialAuth
 
 from weblate.lang.models import Language
-from weblate.trans.models import Project, Change
+from weblate.trans.models import Project, Change, Unit
 from weblate.trans.util import get_site_url, get_distinct_translations
 from weblate.accounts.avatar import get_user_display
 import weblate
@@ -554,7 +554,6 @@ class Profile(models.Model):
         '''
         Returns list of secondary units.
         '''
-        from weblate.trans.models.unit import Unit
         secondary_langs = self.secondary_languages.exclude(
             id=unit.translation.language.id
         )
@@ -663,6 +662,7 @@ def create_groups(update):
             Permission.objects.get(codename='delete_comment'),
             Permission.objects.get(codename='add_suggestion'),
             Permission.objects.get(codename='use_mt'),
+            Permission.objects.get(codename='edit_priority'),
         )
 
     try:
@@ -731,14 +731,14 @@ def sync_create_groups(sender, app, **kwargs):
 
 
 @receiver(post_save, sender=User)
-def create_profile_callback(sender, **kwargs):
+def create_profile_callback(sender, instance, created=False, **kwargs):
     '''
     Automatically adds user to Users group.
     '''
-    if kwargs['created']:
+    if created:
         # Add user to Users group if it exists
         try:
             group = Group.objects.get(name='Users')
-            kwargs['instance'].groups.add(group)
+            instance.groups.add(group)
         except Group.DoesNotExist:
             pass