+++ weblate/trans/models/__init__.py
@@ -21,7 +21,7 @@
 import os
 import shutil
 
-from django.db.models.signals import post_delete
+from django.db.models.signals import post_delete, post_save
 from django.dispatch import receiver
 
 from weblate.trans.models.project import Project
@@ -29,20 +29,22 @@ from weblate.trans.models.subproject import SubProject
 from weblate.trans.models.translation import Translation
 from weblate.trans.models.unit import Unit
 from weblate.trans.models.unitdata import (
-    Check, Suggestion, Comment, IndexUpdate, Vote
+    Check, Suggestion, Comment, Vote
 )
+from weblate.trans.models.search import IndexUpdate
 from weblate.trans.models.changes import Change
 from weblate.trans.models.dictionary import Dictionary
 from weblate.trans.models.source import Source
 from weblate.trans.models.advertisement import Advertisement
+from weblate.trans.models.whiteboard import WhiteboardMessage
 
 
 @receiver(post_delete, sender=Project)
 @receiver(post_delete, sender=SubProject)
 def delete_object_dir(sender, instance, **kwargs):
-    '''
+    """
     Handler to delete (sub)project directory on project deletion.
-    '''
+    """
     # Do not delete linked subprojects
     if hasattr(instance, 'is_repo_link') and instance.is_repo_link:
         return
@@ -52,3 +54,17 @@ def delete_object_dir(sender, instance, **kwargs):
     # Remove path if it exists
     if os.path.exists(project_path):
         shutil.rmtree(project_path)
+
+
+@receiver(post_save, sender=Source)
+def update_string_pririties(sender, instance, created=False, **kwargs):
+    """
+    Updates unit score
+    """
+    if instance.priority_modified:
+        units = Unit.objects.filter(
+            checksum=instance.checksum
+        ).exclude(
+            priority=instance.priority
+        )
+        units.update(priority=instance.priority)