+++ weblate/trans/views/edit.py
@@ -181,11 +181,14 @@ def search(translation, request):
     # Remove old search results
     cleanup_session(request.session)
 
+    if name is not None:
+        name = unicode(name)
+
     # Store in cache and return
     search_id = str(uuid.uuid1())
     search_result = {
         'query': search_query,
-        'name': unicode(name),
+        'name': name,
         'ids': unit_ids,
         'search_id': search_id,
         'ttl': int(time.time()) + 86400,
@@ -314,8 +317,7 @@ def handle_translate(translation, request, user_locked,
             request,
             _('You don\'t have privileges to save translations!')
         )
-    elif (unit.only_vote_suggestions()
-            and not request.user.has_perm('trans.save_translation')):
+    elif unit.only_vote_suggestions():
         messages.error(
             request,
             _('Only suggestions are allowed in this translation!')